result is always integer, no type cast required

+++ src/test/java/com/aif/common/QualityOutputResults.java
@@ -148,7 +148,7 @@ public class QualityOutputResults {
 	}
 	
 	private int compareWithStandard(int number) {
-		return (int) (number * 100) / this.standardSentencesNumber;
+		return (number * 100) / this.standardSentencesNumber;
 	}
 
 }
